FYR-7497: Product carousel should be navigable via keyboard
https://jira.corp.adobe.com/browse/FYR-7497
