+++ src/content/templates/product-carousel.mustache
@@ -5,10 +5,10 @@
 {{/productCarouselTitleShow}}
 <div class="product-carousel-container">
   <div class="product-carousel-left">
-    <div class="product-arrow"></div>
+    <div class="product-arrow" role="button"></div>
   </div>
   <div class="product-carousel-body"></div>
   <div class="product-carousel-right">
-    <div class="product-arrow"></div>
+    <div class="product-arrow" role="button"></div>
   </div>
 </div>
